requirements.txt
CherryPy==18.6.0
Jinja2==2.11.2
PyJWT==1.7.1
cryptography==2.9.2
geoip2==3.0.0
httplib2==0.18.1
mistletoe==0.7.2
pytz==2020.1
requests==2.23.0
sdnotify==0.3.2
